pyproject.toml
[project]
name = "app-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.6",
    "playwright>=1.57.0",
    "python-dotenv>=1.2.1",
]
